% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spina_functions.R
\name{estimated.sGD}
\alias{estimated.sGD}
\title{Calculated standardised step-up deiodinase activity (SPINA-sGD)}
\usage{
estimated.sGD(FT4, FT3, mean = 30, sd = 5)
}
\arguments{
\item{FT4}{free T4 concentration in pmol/L}

\item{FT3}{free T3 concentration in pmol/L}

\item{mean}{mean SPINA-GD (optional, default value 30)}

\item{sd}{standard deviation of SPINA-GD (optional, default value 5)}
}
\value{
Returns standardised SPINA-GD, a calculated biomarker for the sum activity of peripheral step-up deiodinases (z-transformed variant)
}
\description{
This function calculates the sum activity of peripheral step-up-deiodinases
(SPINA-GD) from steady-state concentrations of free T4 and free T3.
SPINA-GD is an inferred representation of GD, the total activity of the type 1
deiodinase under stimulated conditions.
This parameter is an implementation of the structure parameter inference approach (SPINA).
The result is delivered in standardised form (z-transformed).
}
\details{
This function is able to do vectorised calculations. See \link{SPINA.sGD} for a an implementation
with a more modern nomenclature.
}
\note{
The software functions described in this document are intended for research use only.
Hormone concentrations should have been obtained simultaneously in order to
avoid bias by transition effects.
}
\examples{
estimated.sGD(16.5, 4.5)
}
\references{
Dietrich JW, Landgrafe G, Fotiadou EH. TSH and Thyrotropic Agonists: Key Actors in Thyroid Homeostasis. J Thyroid Res. 2012;2012:351864. doi: 10.1155/2012/351864. Epub 2012 Dec 30. PMID: 23365787; PMCID: PMC3544290.

Dietrich JW, Landgrafe-Mende G, Wiora E, Chatzitomaris A, Klein HH, Midgley JE, Hoermann R. Calculated Parameters of Thyroid Homeostasis: Emerging Tools for Differential Diagnosis and Clinical Research. Front Endocrinol (Lausanne). 2016 Jun 9;7:57. doi: 10.3389/fendo.2016.00057. PMID: 27375554; PMCID: PMC4899439.
}
\author{
Johannes W. Dietrich, Eleni Karamitrou
}
