% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spina_functions.R
\name{HOMA.Beta}
\alias{HOMA.Beta}
\title{Homeostasis model assessment: beta-cell function (HOMA-Beta)}
\usage{
HOMA.Beta(Insulin, Glucose)
}
\arguments{
\item{Insulin}{Insulin concentration in pmol/L}

\item{Glucose}{Glucose concentration in mmol/L}
}
\value{
Returns the HOMA-Beta index, a calculated biomarker for beta-cell function
}
\description{
This function calculates the beta-cell function (HOMA-Beta) from steady-state
fasting concentrations of insulin and glucose. HOMA-Beta is one of the parameters
provided by the homeostasis model assessment.
}
\details{
This function is able to do vectorised calculations.
}
\note{
The software functions described in this document are intended for research use only.
Hormone or metabolite concentrations should have been obtained simultaneously in order to
avoid bias by transition effects.
}
\examples{
HOMA.Beta(63.01, 4.34)
}
\author{
Johannes W. Dietrich, Eleni Karamitrou, Bernhard O. Boehm
}
