% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spina_functions.R
\name{SPINA.DI}
\alias{SPINA.DI}
\title{Static disposition index (SPINA-DI)}
\usage{
SPINA.DI(Insulin, Glucose)
}
\arguments{
\item{Insulin}{Insulin concentration in pmol/L}

\item{Glucose}{Glucose concentration in mmol/L}
}
\value{
Returns the static disposition index (SPINA-DI), a
representation of the loop gain of glucose homeostasis as a numeric result representing
a single value or a vector, depending on the vector length of the arguments.
}
\description{
This function calculates the static disposition index (SPINA-DI) from steady-state
fasting concentrations of insulin and glucose. SPINA-DI is an inferred
representation of the loop gain of the homeostatic feedback control system.
This parameter is an implementation of the structure parameter inference approach (SPINA).
}
\details{
This function is able to do vectorised calculations.
}
\note{
The software functions described in this document are intended for research use only.
Hormone or metabolite concentrations should have been obtained simultaneously in order to
avoid bias by transition effects.
}
\examples{
SPINA.DI(63.01, 4.34)
}
\references{
Dietrich JW, Dasgupta R, Anoop S, Jebasingh F, Kurian ME, Inbakumari M, Boehm BO, Thomas N. SPINA Carb: a simple mathematical model supporting fast in-vivo estimation of insulin sensitivity and beta cell function. Sci Rep. 2022 Oct 21;12(1):17659. doi: 10.1038/s41598-022-22531-3. PMID: 36271244; PMCID: PMC9587026.

Dietrich JW, Abood A, Dasgupta R, Anoop S, Jebasingh FK, Spurgeon R, Thomas N, Boehm BO. A novel simple disposition index (SPINA-DI) from fasting insulin and glucose concentration as a robust measure of carbohydrate homeostasis. J Diabetes. 2024 Sep;16(9):e13525. doi: 10.1111/1753-0407.13525. Epub 2024 Jan 2. PMID: 38169110; PMCID: PMC11418405.
}
\author{
Johannes W. Dietrich, Eleni Karamitrou, Bernhard O. Boehm
}
