% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spina_functions.R
\docType{data}
\name{Vellore}
\alias{Vellore}
\title{Data for SPINA Carb}
\format{
\subsection{"Vellore"}{

A data frame with 117 rows and 14 columns
\describe{
\item{Age}{Age in years}
\item{Height}{Body height in cm}
\item{Body.Mass}{Body mass in kg}
\item{BMI}{Body mass index in kg/m^2}
\item{Fasting.Glucose}{Fasting glucose concentration in mg/dL}
\item{Fasting.Insulin}{Fasting insulin concentration in mIU/L}
\item{SPINA_GR}{Insulin receptor gain (SPINA-GR) in mol/s}
\item{SPINA_GBeta}{Secretory capacity of pancreatic beta-cells (SPINA-GBeta) in pmol/s}
\item{SPINA_DI}{Static disposition index (SPINA-DI)}
\item{HOMA_Beta}{Estimate of beta-cell function based on the Homeostasis Model Assessment (HOMA-Beta)}
\item{HOMA_IR}{Estimate of insulin resistance based on the Homeostasis Model Assessment (HOMA-IR)}
\item{HOMA_IS}{Estimate of insulin sensitivity based on the Homeostasis Model Assessment (HOMA-IS)}
\item{QUICKI_Score}{Quantitative Insulin Sensitivity Check Index (QUICKI)}
}
}
}
\source{
Dietrich JW, Dasgupta R, Anoop S, Jebasingh F, Kurian ME, Inbakumari M, Boehm BO,
Thomas N. SPINA Carb: a simple mathematical model supporting fast in-vivo estimation
of insulin sensitivity and beta cell function. Sci Rep. 2022 Oct 21;12(1):17659.
doi 10.1038/s41598-022-22531-3. PMID 36271244; PMCID PMC9587026.

Dietrich JW, Abood A, Dasgupta R, Anoop S, Jebasingh FK, Spurgeon R, Thomas N,
Boehm BO. A novel simple disposition index (SPINA-DI) from fasting insulin and
glucose concentration as a robust measure of carbohydrate homeostasis. J Diabetes.
2024 Jan 2. doi: 10.1111/1753-0407.13525. PMID: 38169110.
}
\description{
Vellore: Data from Dietrich et al. (2022 and 2024)
}
