% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spina_functions.R
\name{estimated.GDTT}
\alias{estimated.GDTT}
\title{Calculated step-up deiodinase activity, based on total T4 and T3 (SPINA-GDTT)}
\usage{
estimated.GDTT(T4, T3)
}
\arguments{
\item{T4}{total T4 concentration in nmol/L}

\item{T3}{tptaö T3 concentration in nmol/L}
}
\value{
Returns SPINA-GD, a calculated biomarker for the sum activity of peripheral step-up deiodinases
}
\description{
Calculated step-up deiodinase activity, based on total T4 and T3 (SPINA-GDTT)
}
\details{
This function is able to do vectorised calculations.
}
\note{
The software functions described in this document are intended for research use only.
Hormone concentrations should have been obtained simultaneously in order to
avoid bias by transition effects.
}
\examples{
estimated.GDTT(90, 2.5)
}
\references{
Dietrich JW, Landgrafe G, Fotiadou EH. TSH and Thyrotropic Agonists: Key Actors in Thyroid Homeostasis. J Thyroid Res. 2012;2012:351864. doi: 10.1155/2012/351864. Epub 2012 Dec 30. PMID: 23365787; PMCID: PMC3544290.

Dietrich JW, Landgrafe-Mende G, Wiora E, Chatzitomaris A, Klein HH, Midgley JE, Hoermann R. Calculated Parameters of Thyroid Homeostasis: Emerging Tools for Differential Diagnosis and Clinical Research. Front Endocrinol (Lausanne). 2016 Jun 9;7:57. doi: 10.3389/fendo.2016.00057. PMID: 27375554; PMCID: PMC4899439.
}
\author{
Johannes W. Dietrich
}
