% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spina_functions.R
\name{estimated.GTT}
\alias{estimated.GTT}
\title{Calculated thyroid's secretory capacity, based on total T4 (SPINA-GTT)}
\usage{
estimated.GTT(TSH, T4)
}
\arguments{
\item{TSH}{thyrotropin concentration in mIU/L}

\item{T4}{total T4 concentration in nmol/L}
}
\value{
Returns SPINA-GTT, a calculated biomarker for thyroid's secretory capacity (aka thyroid output)
}
\description{
This function calculates the secretory capacity of the thyroid gland for T4
(SPINA-GT) from steady-state concentrations of TSH and total T4.
SPINA-GT is an inferred representation of GT, the maximum amount of thyroxine
that can be produced by the thyroid under stimulated conditions.
This parameter is an implementation of the structure parameter inference approach (SPINA).
}
\details{
This function is able to do vectorised calculations. See \link{SPINA.GTT} for a an implementation
with a more modern nomenclature.
}
\note{
The software functions described in this document are intended for research use only.
Hormone concentrations should have been obtained simultaneously in order to
avoid bias by transition effects.
}
\examples{
estimated.GTT(1.0, 90)
}
\references{
Dietrich JW, Landgrafe G, Fotiadou EH. TSH and Thyrotropic Agonists: Key Actors in Thyroid Homeostasis. J Thyroid Res. 2012;2012:351864. doi: 10.1155/2012/351864. Epub 2012 Dec 30. PMID: 23365787; PMCID: PMC3544290.

Dietrich JW, Landgrafe-Mende G, Wiora E, Chatzitomaris A, Klein HH, Midgley JE, Hoermann R. Calculated Parameters of Thyroid Homeostasis: Emerging Tools for Differential Diagnosis and Clinical Research. Front Endocrinol (Lausanne). 2016 Jun 9;7:57. doi: 10.3389/fendo.2016.00057. PMID: 27375554; PMCID: PMC4899439.
}
\author{
Johannes W. Dietrich
}
