% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spina_functions.R
\name{estimated.TTSI}
\alias{estimated.TTSI}
\title{Thyrotroph thyroid hormone sensitivity index (TTSI)}
\usage{
estimated.TTSI(TSH, FT4, lu)
}
\arguments{
\item{TSH}{thyrotropin concentration in mIU/L}

\item{FT4}{free T4 concentration in pmol/L}

\item{lu}{upper limit of the reference interval of FT4 in pmol/L}
}
\value{
Returns the TTSI, a calculated biomarker for the sensitivity of pituitary thyrotrophic cells to thyroxine
}
\description{
This function calculates the thyrotroph thyroid hormone sensitivity index
(TTSI) from steady-state concentrations of TSH and free T4. This parameter is also
referred to as Thyrotroph T4 Resistance Index or TT4RI.
}
\details{
This function is able to do vectorised calculations.
}
\note{
The software functions described in this document are intended for research use only.
Hormone concentrations should have been obtained simultaneously in order to
avoid bias by transition effects.
}
\examples{
estimated.TTSI(1.0, 16.5, 18)
}
\references{
Dietrich JW, Landgrafe G, Fotiadou EH. TSH and Thyrotropic Agonists: Key Actors in Thyroid Homeostasis. J Thyroid Res. 2012;2012:351864. doi: 10.1155/2012/351864. Epub 2012 Dec 30. PMID: 23365787; PMCID: PMC3544290.

Dietrich JW, Landgrafe-Mende G, Wiora E, Chatzitomaris A, Klein HH, Midgley JE, Hoermann R. Calculated Parameters of Thyroid Homeostasis: Emerging Tools for Differential Diagnosis and Clinical Research. Front Endocrinol (Lausanne). 2016 Jun 9;7:57. doi: 10.3389/fendo.2016.00057. PMID: 27375554; PMCID: PMC4899439.

Pohlenz J, Weiss RE, Macchia PE, Pannain S, Lau IT, Ho H, Refetoff S. Five new families with resistance to thyroid hormone not caused by mutations in the thyroid hormone receptor beta gene. J Clin Endocrinol Metab. 1999 Nov;84(11):3919-28. doi: 10.1210/jcem.84.11.6080. PMID: 10566629.
}
\author{
Johannes W. Dietrich
}
