% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spina_functions.R
\name{HOMA.IR}
\alias{HOMA.IR}
\title{Homeostasis model assessment: insulin resistance (HOMA-IR)}
\usage{
HOMA.IR(Insulin, Glucose)
}
\arguments{
\item{Insulin}{Insulin concentration in pmol/L}

\item{Glucose}{Glucose concentration in mmol/L}
}
\value{
Returns the HOMA-IR index, a calculated biomarker for insulin resistance.
}
\description{
This function calculates the insulin resistance index (HOMA-IR) from steady-state
fasting concentrations of insulin and glucose. HOMA-IR is one of the parameters
provided by the homeostasis model assessment.
}
\details{
This function is able to do vectorised calculations.
}
\note{
The software functions described in this document are intended for research use only.
Hormone or metabolite concentrations should have been obtained simultaneously in order to
avoid bias by transition effects.
}
\examples{
HOMA.IR(63.01, 4.34)
}
\author{
Johannes W. Dietrich, Bernhard O. Boehm
}
