% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spina_functions.R
\docType{data}
\name{Pilo}
\alias{Pilo}
\title{Data for SPINA}
\format{
\subsection{"Pilo"}{

A data frame with 14 rows and 33 columns
\describe{
\item{Sex}{Sex (f: female; m: male)}
\item{Age}{Age in years}
\item{Body.Mass}{Body mass in kg}
\item{BSA}{Body surface area in m^2}
\item{IDV.T4}{initial distribution volume of thyroxine in mL}
\item{TT4}{Total T4 concentration in µg/dL}
\item{TT4.SI}{Total T4 concentration in nmol/L}
\item{TT3}{Total T3 concentration in ng/mL}
\item{TT3.SI}{Total T3 concentration in nmol/L}
}
}
}
\source{
Dietrich JW, Landgrafe-Mende G, Wiora E, Chatzitomaris A, Klein HH, Midgley JE, Hoermann R. Calculated Parameters of Thyroid Homeostasis: Emerging Tools for Differential Diagnosis and Clinical Research. Front Endocrinol (Lausanne). 2016 Jun 9;7:57. doi 10.3389/fendo.2016.00057. PMID 27375554; PMCID PMC4899439.

Pilo A, Iervasi G, Vitek F, Ferdeghini M, Cazzuola F, Bianchi R. Thyroidal and
peripheral production of 3,5,3'-triiodothyronine in humans by multicompartmental
analysis. Am J Physiol. 1990 Apr;258(4 Pt 1):E715-26. PMID 2333963.
}
\description{
Pilo: Data from Pilo et al. (1990)
}
