\name{pilo}
\alias{pilo}
\title{Demo Data Set from Pilo et al. 1990}
\description{Demo data set from Pilo et al. 1990.  }
\details{This is a demo data set taken from a compartment-analytical study by Pilo et al. 1990.  }
\value{This table contains the following slots:
  \item{BSA}{body surface area in m^2}
  \item{IDV}{initial distribution volume}
  \item{TT4}{total T4 in mcg/dl}
  \item{TT3}{total T3 in ng/ml}
  \item{FT4}{free T4 in pg/ml}
  \item{FT3}{free T3 in pg/ml}
  \item{TT4.SI}{total T4 in nmol/l}
  \item{TT3.SI}{total T3 in nmol/l}
  \item{FT4.SI}{free T4 in pmol/l}
  \item{FT3.SI}{free T3 in pmol/l}
  \item{SR}{secretion rate}
  \item{CR.F, CR.S and CR.T}{conversion rate (fast pool, slow
pool and total)}
  \item{PAR}{plasma apperance rate}
  \item{PR}{production rate}
  \item{CR.6, CR.2, CR.0}{conversion ratio from 6-compartment,
2-compartment and noncompartmental model}
}
\references{
Pilo A, Iervasi G, Vitek F, Ferdeghini M, Cazzuola F, Bianchi R. Thyroidal and peripheral production of 3,5,3'-triiodothyronine in humans by multicompartmental  analysis. Am J Physiol. 1990 Apr;258(4 Pt 1):E715-26. PMID: 2333963.

Dietrich J. W., Landgrafe-Mende G., Wiora E., Chatzitomaris A., Klein H. H., Midgley J. E., Hoermann R. 2016 Calculated Parameters of Thyroid Homeostasis: Emerging Tools for  Differential Diagnosis and Clinical Research. \emph{Front Endocrinol (Lausanne)}. 2016 Jun 9;7:57. doi: 10.3389/fendo.2016.00057. PMID: 27375554; PMCID: PMC4899439.
}

\seealso{
\code{\link{SPINA.GD}}, \code{\link{estimated.GT}}, \code{\link{SPINA.GT}}, \code{\link{estimated.GDTT}}, \code{\link{SPINA.GDTT}}, \code{\link{estimated.GTT}}, \code{\link{SPINA.GTT}}, \code{\link{estimated.sGD}}, \code{\link{SPINA.sGD}}, \code{\link{estimated.TSHI}}, \code{\link{estimated.sTSHI}}, \code{\link{estimated.TTSI}},
}
\keyword{SPINA}