% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spina_functions.R
\name{QUICKI}
\alias{QUICKI}
\title{Quantitative insulin-sensitivity check index (QUICKI)}
\usage{
QUICKI(Insulin, Glucose)
}
\arguments{
\item{Insulin}{Insulin concentration in pmol/L}

\item{Glucose}{Glucose concentration in mmol/L}
}
\value{
Returns the QUICKI index, a calculated biomarker for insulin sensitivity
}
\description{
This function calculates the quantitative insulin sensitivity check index (QUICKI)
from steady-state fasting concentrations of insulin and glucose.
}
\details{
This function is able to do vectorised calculations.
}
\note{
The software functions described in this document are intended for research use only.
Hormone or metabolite concentrations should have been obtained simultaneously in order to
avoid bias by transition effects.
}
\examples{
QUICKI(63.01, 4.34)
}
\author{
Johannes W. Dietrich, Bernhard O. Boehm
}
